+++ vcflat/HeaderExtraction.py
@@ -1,60 +1,87 @@
 from collections import defaultdict
 from cyvcf2 import VCF
 import pprint as pp
+import sys
 
 
-class VcfHeader(object):
+class VcfHeader:
     """Class around meta information about vcf file as well as functions to view it"""
 
-    def __init__(self, input_vcf, header=None, meta_dict=None):
+    def __init__(self, input_vcf, samples_in_header=None):
 
         self.input_vcf = input_vcf
-        self.header = header if header is not None else []
-        self.meta_dict = meta_dict if meta_dict is not None else {}
+        self.header = self.populatevcfheader(samples_in_header=samples_in_header)
+        self.meta_dict = self.process_meta_dict()
 
-    def pprint_meta(self, key=False):
-        self.meta_dict = process_meta_dict(self.input_vcf)
-        if key:
+    def get_raw_header(self):
+        """
+        Gets input_vcf as input
+        returns the raw header of the vcf file
+        """
+        vcf_file = VCF("{}".format(self.input_vcf), strict_gt=True)
+        raw_vcf_header_list = [
+            header_line
+            for header_line in vcf_file.raw_header.split("\n")
+            if header_line
+        ]
+        return raw_vcf_header_list
 
-            try:
-                print(f'\n'
-                      f'Printing out the values from the meta dict for {key}\n'
-                      f'Starts with a key and then gives a info for the output\n')
-                for k, v in self.meta_dict[key].items():
-                    print(f"Key = {k} : Values = {v}")
-            except AttributeError:
-                print(f'Seems like the key value pairs did not play nicely for {key} \n'
-                      'but here is the output anyway ')
-                print(self.meta_dict[key])
-        else:
-            pp.pprint(self.meta_dict, depth=3)
+    def extract_header(self):
+        """
+        looks for the real header file of the vcf file and creates a vcf_header list
+        :returns header_list
+        """
+        vcf_header = ""
+        raw_header = self.get_raw_header()
+        for i in raw_header:
+            if i.startswith("#CHROM"):
+                vcf_header = [ii for ii in i.split("\t")]
+        return vcf_header
 
-    def determine_keys(self):
-        pass
+    def process_meta_dict(self):
+        """
+        main parsing of the vcf header (everything that starts with ## into a nice dict of all meta information
+        :param inputvcf: basic VCF file
+        :return: a nice dict with all header elements and their subelements structured
+        """
+        raw_header = self.get_raw_header()
+        raw_header_popped = pop_header(raw_header)
+        cleaned_meta = clean_meta(raw_header_popped)
+        base_dict = dictify(cleaned_meta)
+
+        chunk_dict = validate_meta(base_dict)
+
+        meta_dict = {}
+        if chunk_dict["INFO"]:
+            meta_dict = generate_complete_dict(base_dict, "INFO", 3)
+        if chunk_dict["FORMAT"]:
+            meta_dict = generate_complete_dict(base_dict, "FORMAT", 3)
+        if chunk_dict["FILTER"]:
+            meta_dict = generate_complete_dict(meta_dict, "FILTER", 1)
+
+        meta_dict = detect_double_type(meta_dict)
+        return meta_dict
+
+    def populatevcfheader(self, samples_in_header=None):
+        """
+        returns a header on the original vcf with "samples_in_header" being the sample names (fields after the "INFO" samples)
 
 
-def get_raw_header(input_vcf):
-    """
-    Gets input_vcf as input
-    returns the raw header of the vcf file
-    """
-    vcf_file = VCF('{}'.format(input_vcf), strict_gt=True)
-    raw_vcf_header_list = [header_line for header_line in vcf_file.raw_header.split("\n") if header_line]
-
-    return raw_vcf_header_list
-
-
-def extract_header(input_vcf):
-    """
-    looks for the real header file of the vcf file and creates a vcf_header list
-    :returns header_list
-    """
-    vcf_header = ''
-    raw_header = get_raw_header(input_vcf)
-    for i in raw_header:
-        if i.startswith('#CHROM'):
-            vcf_header = [ii for ii in i.split('\t')]
-    return vcf_header
+        :param input_vcf: basic VCF file
+        :param samples_in_header: samplenames
+        :return: things that should be in the actual header of the vcf
+        """
+        header = self.extract_header()
+        if samples_in_header:
+            samples_in_header = samples_in_header.split()
+            if len(samples_in_header) == len(header[9:]):
+                header = header[:9] + samples_in_header
+            else:
+                sys.exit(
+                    f" '--samples_in_header' given has {len(samples_in_header)}, "
+                    f"but there are {len(header[9:])} samples columns in the vcf body header"
+                )
+        return header
 
 
 def pop_header(raw_header):
@@ -63,7 +90,7 @@ def pop_header(raw_header):
     :return:
     """
     for i in raw_header:
-        if i.startswith('#CHROM'):
+        if i.startswith("#CHROM"):
             raw_header.pop()
     return raw_header
 
@@ -75,13 +102,13 @@ def clean_meta(unclean_list):
     """
     clean_list = []
     for i in unclean_list:
-        if '=<' in i:
-            i = i.rstrip('>')
-            i = i.replace('##', '')
-            ii = i.split('=<', 1)
+        if "=<" in i:
+            i = i.rstrip(">")
+            i = i.replace("##", "")
+            ii = i.split("=<", 1)
         else:
-            i = i.replace('##', '')
-            ii = i.split('=', 1)
+            i = i.replace("##", "")
+            ii = i.split("=", 1)
         clean_list.append(ii)
     return clean_list
 
@@ -101,7 +128,7 @@ def dictify(clean_list):
             else:
                 base_dict[key].append(value)
         except IndexError:
-            print('index error')
+            print("index error")
 
     return base_dict
 
@@ -112,12 +139,14 @@ def validate_meta(base_dict):
     :return:
     """
     metalist = [i for i in base_dict]
-    includelist = ['INFO', 'FORMAT', 'FILTER']
+    includelist = ["INFO", "FORMAT", "FILTER"]
     inmeta = {}
     for i in includelist:
-        if i in metalist:
-            inmeta[i] = True
-        else:
+        try:
+            if i in metalist:
+                inmeta[i] = True
+        except:
+            # TODO create a breaking error here (not forget to fix test)
             inmeta[i] = False
     return inmeta
 
@@ -134,7 +163,7 @@ def custom_dict_chunking(basedict, field, how_many):
 
     new_dict_list = []
     for v in basedict[field]:
-        new_dict_list.append(v.split(',', how_many))
+        new_dict_list.append(v.split(",", how_many))
     basedict[field] = new_dict_list
 
     return basedict
@@ -155,7 +184,7 @@ def list2dict(basedict, field):
         l4d = []
         for i in info_list:
             l4d.append(i.split("=", 1)[1])
-        basedict[field][l4d[0]] = l4d[1:]
+        basedict[field][l4d[0]] = {"data": l4d[1:]}
     return basedict
 
 
@@ -174,41 +203,12 @@ def generate_complete_dict(basedict, field, how_many):
     return meta_dict
 
 
-def process_meta_dict(inputvcf):
-    """
-    main parsing of the vcf header (everything that starts with ## into a nice dict of all meta information
-    :param inputvcf: basic VCF file
-    :return: a nice dict with all header elements and their subelements structured
-    """
-    raw_header = get_raw_header(inputvcf)
-    raw_header_popped = pop_header(raw_header)
-    cleaned_meta = clean_meta(raw_header_popped)
-    base_dict = dictify(cleaned_meta)
-
-    chunk_dict = validate_meta(base_dict)
-
-    meta_dict = {}
-    if chunk_dict['INFO']:
-        meta_dict = generate_complete_dict(base_dict, "INFO", 3)
-    if chunk_dict['FORMAT']:
-        meta_dict = generate_complete_dict(base_dict, 'FORMAT', 3)
-    if chunk_dict['FILTER']:
-        meta_dict = generate_complete_dict(meta_dict, 'FILTER', 1)
-    return meta_dict
-
-
-def populatevcfheader(input_vcf, samplefield=None):
-    """
-    returns a header on the original vcf with "samplefield" being the sample names (fields after the "INFO" samples)
-
-
-    :param input_vcf: basic VCF file
-    :param samplefield: samplenames
-    :return: things that should be in the actual header of the vcf
-    """
-    metadict = process_meta_dict(input_vcf)
-    header = extract_header(input_vcf)
-    if samplefield:
-        header = header[:9] + [samplefield]
-    vcf_header = VcfHeader(input_vcf, header, metadict)
-    return vcf_header
+def detect_double_type(metadict):
+    for k, v in metadict["FORMAT"].items():
+        double_type = "NAN"
+        if any("ref" and "alt" in i for i in v["data"]):
+            double_type = "REF_ALT"
+        if any("tier" in i for i in v["data"]):
+            double_type = "TIERS"
+        metadict["FORMAT"][k]["double_type"] = double_type
+    return metadict