+++ vcflat/tests/test_HeaderExtraction/test_detect_double_type.py
@@ -1,5 +1,4 @@
 import os
-
 from vcflat.HeaderExtraction import VcfHeader
 
 
@@ -7,15 +6,9 @@ def get_input():
     test_data_dir = os.path.join(os.path.dirname(__file__), "..", "test_data")
     i = os.path.join(test_data_dir, "test.snpeff.vcf")
     vcfh = VcfHeader(i)
-    output = vcfh.extract_header()
-    return output
+    return vcfh
 
 
 def test_1():
-    """ checks that output is list """
-    assert type(get_input()) is list
-
-
-def test_2():
-    """ checks that output is not empty """
-    assert len(get_input()) != 0
+    dtd = get_input()
+    assert dtd.meta_dict["FORMAT"]["AD"]["double_type"] == "REF_ALT"