+++ vcflat/tests/test_HeaderExtraction/test_get_raw_header.py
@@ -1,16 +1,20 @@
 import os
-from vcflat.HeaderExtraction import  get_raw_header
+from vcflat.HeaderExtraction import VcfHeader
+
 
 def get_input():
-    test_data_dir = os.path.join(os.path.dirname(__file__), '..', 'test_data')
+    test_data_dir = os.path.join(os.path.dirname(__file__), "..", "test_data")
     i = os.path.join(test_data_dir, "test.snpeff.vcf")
-    output = get_raw_header(i)
+    vcfh = VcfHeader(i)
+    output = vcfh.get_raw_header()
     return output
+
 
 def test_1():
     """ checks that output is list """
     assert type(get_input()) is list
 
+
 def test_2():
     """ checks that output is not empty """
-    assert len( get_input()) != 0
+    assert len(get_input()) != 0