+++ vcflat/tests/test_VcfParse/test_get_header_fast.py
@@ -3,7 +3,7 @@ import os
 
 
 def get_input():
-    test_data_dir = os.path.join(os.path.dirname(__file__), '..', 'test_data')
+    test_data_dir = os.path.join(os.path.dirname(__file__), "..", "test_data")
     i = os.path.join(test_data_dir, "test-hemi.vcf")
     vcffile = VP.VcfParse(i)
     return vcffile