+++ vcflat/tests/test_VcfParse/test_nestlists.py
@@ -2,7 +2,9 @@ import vcflat.VcfParse as VP
 
 
 def base_tests():
-    return VP.nestlists(['','','','','','','','','GT:AD:DP:GQ:PL','0/0:2,0:2:6.01:0,6,6'])
+    return VP.nestlists(
+        ["", "", "", "", "", "", "", "", "GT:AD:DP:GQ:PL", "0/0:2,0:2:6.01:0,6,6"]
+    )
 
 
 def test_1():